:white_check_mark: Updated parser tests to check for full date
Before we checked only the year, but it wasn't enough.
Now the tests check for the full string given by `toISOString()`. This is great but it means that running the tests in machines with different timezones will fail.
Because of this I had to add a fixed timezone to run the tests with. (See package.json)